requirements.txt
fastapi>=0.115.12
hdfs>=2.7.3
kafka-python>=2.2.6
minio>=7.2.15
pandas>=2.2.3
plotly>=6.0.1
prefect>=3.4.1
pyspark>=3.5.5
setuptools>=80.4.0
uvicorn>=0.34.2